refactor: follow new file structure
