+++ lib/segment/src/index/field_index/full_text_index/read_only_text_index/read_ops.rs
@@ -3,41 +3,16 @@ use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
 use common::universal_io::{UniversalRead, UserData};
 
-use super::inverted_index::{InvertedIndex, ParsedQuery, TokenId};
-use super::mmap_text_index::MmapFullTextIndex;
-use super::read_ops::{self, FullTextIndexRead};
-use super::tokenizers::Tokenizer;
+use super::super::inverted_index::{InvertedIndex, ParsedQuery, TokenId};
+use super::super::read_ops::{self, FullTextIndexRead};
+use super::super::tokenizers::Tokenizer;
+use super::ReadOnlyFullTextIndex;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::{
     CardinalityEstimation, PayloadBlockCondition, PayloadFieldIndexRead,
 };
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
 use crate::types::{FieldCondition, PayloadKeyType};
-
-/// Read-only counterpart of [`MutableFullTextIndex`][1] / [`ImmutableFullTextIndex`][2].
-///
-/// Thin wrapper over an already-immutable [`MmapFullTextIndex<S>`][3] backing
-/// format, parameterised by a [`UniversalRead`] storage. The
-/// [`PayloadFieldIndexRead`] body (filter / cardinality / payload blocks /
-/// condition checker) is shared with the writable variant through the
-/// [`FullTextIndexRead`] trait and the free functions in [`read_ops`][4]; the
-/// wrapper just re-tags telemetry and surfaces the `S`-generic version of the
-/// same data on the read-only side of `ReadOnlyFieldIndex`.
-///
-/// Not yet constructible — lifecycle (open / build) lands in the follow-up PR
-/// that wires up the rest of [`ReadOnlyFieldIndex`][5], matching the dead-code
-/// state of `ReadOnlyNullIndex` / `ReadOnlyBoolIndex` / `ReadOnlyMapIndex` /
-/// `ReadOnlyGeoMapIndex`.
-///
-/// [1]: super::mutable_text_index::MutableFullTextIndex
-/// [2]: super::immutable_text_index::ImmutableFullTextIndex
-/// [3]: super::mmap_text_index::MmapFullTextIndex
-/// [4]: super::read_ops
-/// [5]: crate::index::field_index::field_index_base::read_only::ReadOnlyFieldIndex
-pub struct ReadOnlyFullTextIndex<S: UniversalRead> {
-    #[allow(dead_code)]
-    inner: MmapFullTextIndex<S>,
-}
 
 impl<S: UniversalRead> FullTextIndexRead for ReadOnlyFullTextIndex<S> {
     fn tokenizer(&self) -> &Tokenizer {
